refactor: move cel environment creation to a subpackage
